LOGBROKER-10406 Move kafka files to lib
