+++ ydb/core/kafka_proxy/actors/kafka_offset_fetch_actor.h
@@ -1,4 +1,5 @@
 #include "actors.h"
+#include <ydb/core/kafka_proxy/kafka_consumer_protocol.h>
 #include <ydb/core/kafka_proxy/kafka_events.h>
 #include "../kqp_helper.h"
