+++ ydb/core/kafka_proxy/actors/kafka_read_session_utils.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "actors.h"
+#include <ydb/core/kafka_proxy/kafka_consumer_protocol.h>
 
 namespace NKafka {
