+++ ydb/core/kafka_proxy/ya.make
@@ -35,15 +35,9 @@ SRCS(
     kafka_connection.h
     kafka_constants.h
     kafka_listener.h
-    kafka.h
-    kafka_log.h
-    kafka_log_impl.h
     kafka_messages.cpp
     kafka_messages.h
-    kafka_messages_int.cpp
-    kafka_messages_int.h
     kafka_proxy.h
-    kafka_records.cpp
     kafka_consumer_protocol.cpp
     kafka_metrics.cpp
     kqp_helper.cpp
@@ -55,10 +49,10 @@ SRCS(
     kafka_producer_instance_id.h
 )
 
-GENERATE_ENUM_SERIALIZATION(kafka.h)
 GENERATE_ENUM_SERIALIZATION(actors/actors.h)
 
 PEERDIR(
+    ydb/library/kafka
     ydb/public/sdk/cpp/src/client/params
     ydb/library/actors/core
     ydb/library/actors/protos
