+++ ydb/library/kafka/kafka_messages_int.h
@@ -11,7 +11,6 @@
 #include <type_traits>
 
 #include "kafka_records.h"
-#include "kafka_consumer_protocol.h"
 #include "kafka_log_impl.h"
 
 namespace NKafka {
